Move version.h to the shared src so it can be used in the native loader
