+++ shared/src/Datadog.Trace.ClrProfiler.Native/Datadog.Trace.ClrProfiler.Native.vcxproj
@@ -315,6 +315,7 @@
     <ClInclude Include="..\..\..\shared\src\native-src\dynamic_library_base.h" />
     <ClInclude Include="..\..\..\shared\src\native-src\miniutf.hpp" />
     <ClInclude Include="..\..\..\shared\src\native-src\miniutfdata.h" />
+    <ClInclude Include="..\..\..\shared\src\native-src\version.h" />
     <ClInclude Include="cor_profiler_class_factory.h" />
     <ClInclude Include="cor_profiler.h" />
     <ClInclude Include="dynamic_instance.h" />
