+++ shared/src/native-src/version.h
