+++ tracer/src/Datadog.Tracer.Native/Datadog.Tracer.Native.vcxproj
@@ -274,6 +274,7 @@
     <ClInclude Include="..\..\..\shared\src\native-src\miniutfdata.h" />
     <ClInclude Include="..\..\..\shared\src\native-src\pal.h" />
     <ClInclude Include="..\..\..\shared\src\native-src\string.h" />
+    <ClInclude Include="..\..\..\shared\src\native-src\version.h" />
     <ClInclude Include="fault_tolerant_cor_profiler_function_control.h" />
     <ClInclude Include="fault_tolerant_tokens.h" />
     <ClInclude Include="fault_tolerant_envionrment_variables_util.h" />
@@ -334,7 +335,6 @@
     <ClInclude Include="stats.h" />
     <ClInclude Include="Synchronized.hpp" />
     <ClInclude Include="tracer_tokens.h" />
-    <ClInclude Include="version.h" />
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="..\..\..\shared\src\native-src\miniutf.cpp" />
